backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.11.0
email-validator>=2.3.0
python-multipart>=0.0.22
aiofiles>=25.1.0
celery[redis]>=5.3.0
redis>=5.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.13
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
spacy>=3.7.0
google-genai>=0.1.0
requests>=2.31.0
meilisearch>=0.31.0
